// src/api/ant/interfaces/ant-responses.interfaces.ts
export interface AntDataByPlateResponse {
    code: number;
    message: string;
    vehicle: Partial<AntVehicle>;
}

export interface AntVehicle {
    activo: string;
    ambitoOper: string;
    anio: string;
    anioMatriculado: string;
    apellido1: string;
    apellido2: string;
    avaluoComercial: string;
    cambioPropietario: string;
    canvcp: string;
    capacidad: string;
    carroceria: string;
    casaComercial: string;
    cedulaPropAnterior: string;
    celular: string;
    chasis: string;
    cilindraje: string;
    claseServicio: string;
    claseTran: string;
    color1: string;
    color2: string;
    color3: string;
    combustible: string;
    comprador: string;
    contrato: string;
    cooperativa: string;
    correo: string;
    desdePir: string;
    desdeSoat: string;
    direccion: string;
    disco: string;
    docPropietario: string;
    estado: string;
    estadoCon: string;
    estadoInf: string;
    estadoOper: string;
    estadoSer: string;
    estadoSoat: string;

    facturaComercial: string;
    fechaCaducidad: string;
    fechaCompra: string;
    fechaCompraVenta: string;
    fechaFinCon: string;
    fechaInicCon: string;
    finVig: string;
    hastaPir: string;
    hastaSoat: string;
    idAlterno: string;
    idAlternoMov: string;
    identBenef: string;
    identPotencialProp: string;
    indor: string;
    infrac: string;
    inicioPir: string;
    inicioSoat: string;
    inicioVig: string;
    institucionRevinca: string;
    kilometraje: string;
    marca: string;
    marcaDesc: string;
    modelo: string;
    modeloDesc: string;
    motor: string;
    nombreBenef: string;
    nombrePotencialProp: string;
    nombrePropAnterior: string;
    numEjes: string;
    numRuedas: string;
    numeroRevisado: string;
    numeroTraspaso: string;
    pais: string;
    placaActual: string;
    placaAnterior: string;
    prendaComercial: string;
    prendaIndustrial: string;
    prohibidoEnajenar: string;

    propietario: string;
    registroPcir: string;
    registroSoat: string;
    remarcadoChasis: string;
    remarcadoMotor: string;
    reservaDominio: string;
    residencia: string;
    resultado: AntResultado;
    robado: string;
    rucCooperativa: string;
    subClaseTran: string;
    telefono: string;
    tipoIdent: string;
    tipoIdentBenef: string;
    tipoIdentPotencialProp: string;
    tipoPes: string;
    tipoServicio: string;
    tipoTran: string;
    tipoVehiculo: string;
    tonelaje: string;
}

export interface AntResultado {
    codError: string;
    error: string;
    exito: string;
    mensaje: string;
}